+++ stablehlo/transforms/ChloLegalizeToStablehlo.cpp
@@ -49,7 +49,7 @@
 #include "stablehlo/dialect/ChloOps.h"
 #include "stablehlo/dialect/StablehloOps.h"
 #include "stablehlo/transforms/Passes.h"
-#include "stablehlo/transforms/passes_utils.h"
+#include "stablehlo/transforms/PassesUtils.h"
 
 namespace mlir {
 namespace stablehlo {
